Reduce field size.
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/server/trunk@18326 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ include/object.h
@@ -248,7 +248,7 @@ typedef struct obj {
     uint8       run_away;          /**< Monster runs away if it's hp goes below this percentage. */
     struct treasureliststruct *randomitems; /**< Items to be generated */
     struct obj  *chosen_skill;  /**< The skill chosen to use */
-    uint32      hide;           /**< The object is hidden, not invisible */
+    uint8      hide;           /**< The object is hidden, not invisible */
     /* changes made by kholland@sunlab.cit.cornell.edu */
     /* allows different movement patterns for attackers */
     sint32      move_status;    /**< What stage in attack mode */
